Revert to ver. 1 manifest, corrected behavior
Reverted back to the ver. 1 manifest because the background page was
not loading correctly and I could not make the ver. 2 manifest means of
accomplishing this work.

Corrected the javascript so that it no longer conflicted with a
mis-documented keyboard shortcut for opening the Inspector in Chrome on
OS X.

Developed a thorough annoyance with how much effort it took to
re-create a perfectly good feature that Google removed.  Hope to fix
this in a later patch.